finished base api

+++ EduNote.API.EF/Models/Answer.cs
@@ -6,13 +6,12 @@ namespace EduNote.API.EF.Models
     {
         public string Body { get; set; }
 
-        public int QuestionId { get; set; }
+        public long QuestionId { get; set; }
 
         public Question Question { get; set; }
-
-
-        public int CreatedById { get; set; }
-        public int? ModifiedById { get; set; }
+        
+        public long CreatedById { get; set; }
+        public long? ModifiedById { get; set; }
 
         public User CreatedBy { get; set; }
         public User ModifiedBy { get; set; }
