+++ EduNote.API.EF/Models/Note.cs
@@ -7,14 +7,14 @@ namespace EduNote.API.EF.Models
         public string Title { get; set; }
         public string Body { get; set; }
 
-        public int SectionId { get; set; }
+        public long SectionId { get; set; }
 
         public Section Section { get; set; }
         public ICollection<NoteTags> NoteTags { get; set; }
 
 
-        public int CreatedById { get; set; }
-        public int? ModifiedById { get; set; }
+        public long CreatedById { get; set; }
+        public long? ModifiedById { get; set; }
 
         public User CreatedBy { get; set; }
         public User ModifiedBy { get; set; }
