+++ EduNote.API.EF/Models/NoteTags.cs
@@ -2,8 +2,8 @@
 {
     public partial class NoteTags : BaseModel
     {
-        public int NoteId { get; set; }
-        public int TagId { get; set; }
+        public long NoteId { get; set; }
+        public long TagId { get; set; }
 
         public Note Note { get; set; }
         public Tag Tag { get; set; }
