+++ EduNote.API.EF/Models/QuestionTags.cs
@@ -2,8 +2,8 @@
 {
     public partial class QuestionTags : BaseModel
     {
-        public int QuestionId { get; set; }
-        public int TagId { get; set; }
+        public long QuestionId { get; set; }
+        public long TagId { get; set; }
 
         public Question Question { get; set; }
         public Tag Tag { get; set; }
