+++ EduNote.API.EF/Models/UserGroups.cs
@@ -2,8 +2,8 @@
 {
     public partial class UserGroups : BaseModel
     {
-        public int UserId { get; set; }
-        public int GroupId { get; set; }
+        public long UserId { get; set; }
+        public long GroupId { get; set; }
 
         public User User { get; set; }
         public Group Group { get; set; }
